fixed timestamp & mutex (un)locking

+++ philo_one/srcs/utils.h
@@ -3,6 +3,6 @@
 
 int		ft_atoi(const char *str);
 size_t	ft_strlen(const char *s);
-void	ft_putnbr(int n);
+void	ft_putnbr(unsigned long n);
 
 #endif
